requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
cssselect2==0.7.0
DateTime==5.4
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flask_weasyprint==1.1.0
fonttools==4.49.0
greenlet==3.0.3
html5lib==1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.0
MarkupSafe==2.1.3
pdfkit==1.0.0
pillow==10.2.0
pycparser==2.21
pydyf==0.9.0
pyphen==0.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.25
tinycss2==1.2.1
typing_extensions==4.9.0
weasyprint==61.1
webencodings==0.5.1
Werkzeug==3.0.1
zope.interface==6.1
zopfli==0.2.3